+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog] and this project adheres to
 ## [Unreleased]
 ### Added
 - Variants of `Class::new_instance` that are `unsafe` or take arguments
+- `PartialEq<[A]>` implementation for `Array<O>` where `O: PartialEq<A>`
 
 ### Fixed
 - Safety of `Class::new_instance`